+++ specs/src/main/java/org/seedstack/business/assembler/dsl/AssembleMultipleWithQualifier.java
@@ -23,6 +23,18 @@ public interface AssembleMultipleWithQualifier extends AssembleMultiple {
      */
     AssembleMultiple with(Annotation qualifier);
 
+    
+    /**
+     * Allows to specify the qualifier of the assembler to use with Named annotation.
+     *
+     * @param qualifier
+     *            the value of javax.inject.Named annotation.
+     * @return the next element of the DSL.
+     */
+    AssembleMultiple with(String qualifier);
+    
+    
+    
     /**
      * Allows to specify the qualifier class of the assembler to use.
      *