+++ specs/src/main/java/org/seedstack/business/domain/DomainEventHandler.java
@@ -25,4 +25,5 @@ public interface DomainEventHandler<E extends DomainEvent> {
     void onEvent(E event);
 
     Class<E> getEventClass();
+        
 }